+++ blues/moves.py
@@ -9,34 +9,12 @@ Authors: Samuel C. Gill
 
 Contributors: Nathan M. Lim, Kalistyn Burley, David L. Mobley
 """
-from simtk import openmm, unit
-from openmmtools import alchemy
-from openmmtools.states import SamplerState, ThermodynamicState, CompoundThermodynamicState
-from blues.integrators import AlchemicalExternalLangevinIntegrator
-from openmmtools.mcmc import LangevinDynamicsMove, MCMCMove
-from blues.simulation import SystemFactory
-from blues import utils
 
-import parmed
+
 import mdtraj
 import numpy
 import abc
-import copy
-import math
-import logging
-import os
-import time
-
-import copy
-import math
-import random
-import sys
 import traceback
-
-import mdtraj
-import numpy
-import parmed
-from simtk import unit
 
 try:
     import openeye.oechem as oechem
@@ -50,402 +28,11 @@ except ImportError:
     print('ImportError: Could not import openeye-toolkits. SideChainMove class will be unavailable.')
 
 
-class NCMCMove(MCMCMove):
-    """This is the base Move class. Move provides methods for calculating properties
-    and applying the move on the set of atoms being perturbed in the NCMC simulation.
-    """
 
-    def __init__(self, timestep, n_steps, atom_subset=None, context_cache=None, reporters=[]):
-        self.timestep = timestep
-        self.n_steps = n_steps
-        self.n_accepted = 0
-        self.n_proposed = 0
-        self.logp_accept = 0
-        self.initial_energy = None
-        self.initial_positions = None
-        self.final_energy = None
-        self.final_positions = None
-        self.proposed_positions = None
-        self.atom_subset = atom_subset
-        self.context_cache = context_cache
-        self.currentStep = 0
-        self.reporters = reporters
-
-    @property
-    def statistics(self):
-        """The acceptance statistics as a dictionary."""
-        return dict(
-            n_accepted=self.n_accepted,
-            n_proposed=self.n_proposed,
-            initial_energy=self.initial_energy,
-            initial_positions=self.initial_positions,
-            final_energy=self.final_energy,
-            proposed_positions=self.proposed_positions,
-            final_positions=self.final_positions,
-            logp_accept=self.logp_accept)
-
-    @statistics.setter
-    def statistics(self, value):
-        self.n_accepted = value['n_accepted']
-        self.n_proposed = value['n_proposed']
-        self.initial_energy = value['initial_energy']
-        self.initial_positions = value['initial_positions']
-        self.final_energy = value['final_energy']
-        self.proposed_positions = value['proposed_positions']
-        self.final_positions = value['final_positions']
-        self.logp_accept = value['logp_accept']
-
-    def _before_integration(self, context, thermodynamic_state):
-        """Execute code after Context creation and before integration."""
-        context_state = context.getState(getPositions=True,
-                                         getVelocities=True,
-                                         getEnergy=True,
-                                         enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-        self.initial_positions = context_state.getPositions()
-        self.initial_energy = thermodynamic_state.reduced_potential(context)
-
-    def _after_integration(self, context, thermodynamic_state):
-        """Execute code after integration.
-
-        After this point there are no guarantees that the Context will still
-        exist, together with its bound integrator and system.
-        """
-        context_state = context.getState(getPositions=True,
-                                         getVelocities=True,
-                                         getEnergy=True,
-                                         enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-
-        self.final_positions = context_state.getPositions()
-        self.final_energy = thermodynamic_state.reduced_potential(context)
-        self.logp_accept = context._integrator.getLogAcceptanceProbability(context)
-
-    def _get_integrator(self, thermodynamic_state):
-        return AlchemicalExternalLangevinIntegrator(
-            alchemical_functions={
-                'lambda_sterics':
-                'min(1, (1/0.3)*abs(lambda-0.5))',
-                'lambda_electrostatics':
-                'step(0.2-lambda) - 1/0.2*lambda*step(0.2-lambda) + 1/0.2*(lambda-0.8)*step(lambda-0.8)'
-            },
-            splitting="H V R O R V H",
-            temperature=thermodynamic_state.temperature,
-            nsteps_neq=self.n_steps,
-            timestep=1.0 * unit.femtoseconds,
-            nprop=1,
-            prop_lambda=0.3)
-
-    def apply(self, thermodynamic_state, sampler_state):
-        """Apply a metropolized move to the sampler state.
-
-        Total number of acceptances and proposed move are updated.
-
-        Parameters
-        ----------
-        thermodynamic_state : openmmtools.states.ThermodynamicState
-           The thermodynamic state to use to apply the move.
-        sampler_state : openmmtools.states.SamplerState
-           The initial sampler state to apply the move to. This is modified.
-
-        """
-        # Check if we have to use the global cache.
-        if self.context_cache is None:
-            context_cache = cache.global_context_cache
-        else:
-            context_cache = self.context_cache
-
-        # Create integrator
-        integrator = self._get_integrator(thermodynamic_state)
-
-        # Create context
-        context, integrator = context_cache.get_context(thermodynamic_state, integrator)
-
-        # Compute initial energy. We don't need to set velocities to compute the potential.
-        # TODO assume sampler_state.potential_energy is the correct potential if not None?
-        sampler_state.apply_to_context(context, ignore_velocities=True)
-        context.setVelocitiesToTemperature(thermodynamic_state.temperature)
-
-        self._before_integration(context, thermodynamic_state)
-
-        try:
-            nextReport = [None]*len(self.reporters)
-            endStep = self.currentStep + self.n_steps
-            while self.currentStep < endStep:
-                nextSteps = endStep-self.currentStep
-                anyReport = False
-                for i, reporter in enumerate(self.reporters):
-                    nextReport[i] = reporter.describeNextReport(self)
-                    if nextReport[i][0] > 0 and nextReport[i][0] <= nextSteps:
-                        nextSteps = nextReport[i][0]
-                        anyReport = True
-
-                stepsToGo = nextSteps
-                while stepsToGo > 10:
-                    integrator.step(10)
-                    stepsToGo -= 10
-                integrator.step(stepsToGo)
-                self.currentStep += stepsToGo
-
-                alch_lambda = integrator.getGlobalVariableByName('lambda')
-                if alch_lambda == 0.5:
-                    # Propose perturbed positions. Modifying the reference changes the sampler state.
-                    #print('Proposing move at lambda = %s...' %alch_lambda)
-                    sampler_state.update_from_context(context)
-                    proposed_positions = self._propose_positions(sampler_state.positions[self.atom_subset])
-                    # Compute the energy of the proposed positions.
-                    sampler_state.positions[self.atom_subset] = proposed_positions
-                    sampler_state.apply_to_context(context)
-
-                if anyReport:
-                    context_state = context.getState(
-                         getPositions=True,
-                         getVelocities=True,
-                         getEnergy=True,
-                         enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-                    context_state.currentStep = self.currentStep
-                    context_state.system = thermodynamic_state.get_system()
-
-                    reports = []
-                    for reporter, report in zip(self.reporters, nextReport):
-                        reports.append((reporter,report))
-                    for reporter, next in reports:
-                         reporter.report(context_state, integrator)
-
-
-        except Exception as e:
-            print(e)
-            # Catches particle positions becoming nan during integration.
-            restart = True
-        else:
-            context_state = context.getState(
-                getPositions=True,
-                getVelocities=True,
-                getEnergy=True,
-                enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-        self._after_integration(context, thermodynamic_state)
-        # Update everything but the collective variables from the State object
-        sampler_state.update_from_context(
-            context_state, ignore_collective_variables=True, ignore_velocities=True)
-        # Update only the collective variables from the Context
-        sampler_state.update_from_context(
-            context, ignore_positions=True, ignore_velocities=True, ignore_collective_variables=False)
-
-    @abc.abstractmethod
-    def _propose_positions(self, positions):
-        """Return new proposed positions.
-
-        These method must be implemented in subclasses.
-
-        Parameters
-        ----------
-        positions : nx3 numpy.ndarray
-            The original positions of the subset of atoms that these move
-            applied to.
-
-        Returns
-        -------
-        proposed_positions : nx3 numpy.ndarray
-            The new proposed positions.
-
-        """
-        pass
-
-class ReportLangevinDynamicsMove(LangevinDynamicsMove):
-
-    def __init__(self,
-        timestep=1.0*unit.femtosecond,
-        collision_rate=10.0/unit.picoseconds,
-        n_steps=1000,
-        reassign_velocities=False,
-        reporters=None,
-        **kwargs):
-        super(ReportLangevinDynamicsMove, self).__init__(self,**kwargs)
-        self.n_steps = n_steps
-        self.resassign_velocities = reassign_velocities
-        self.timestep = timestep
-        self.collision_rate = collision_rate
-        self.currentStep = 0
-        self.reporters = reporters
-
-    def _before_integration(self, context, thermodynamic_state):
-        """Execute code after Context creation and before integration."""
-        context_state = context.getState(getPositions=True,
-                                         getVelocities=True,
-                                         getEnergy=True,
-                                         enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-        self.initial_positions = context_state.getPositions()
-        self.initial_energy = thermodynamic_state.reduced_potential(context)
-        self._usesPBC = thermodynamic_state.is_periodic
-
-
-    def _after_integration(self, context, thermodynamic_state):
-        """Execute code after integration.
-
-        After this point there are no guarantees that the Context will still
-        exist, together with its bound integrator and system.
-        """
-        context_state = context.getState(getPositions=True,
-                                         getVelocities=True,
-                                         getEnergy=True,
-                                         enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-
-        self.final_positions = context_state.getPositions()
-        self.final_energy = thermodynamic_state.reduced_potential(context)
-
-    def apply(self, thermodynamic_state, sampler_state):
-        """Propagate the state through the integrator.
-
-        This updates the SamplerState after the integration. It also logs
-        benchmarking information through the utils.Timer class.
-
-        Parameters
-        ----------
-        thermodynamic_state : openmmtools.states.ThermodynamicState
-           The thermodynamic state to use to propagate dynamics.
-        sampler_state : openmmtools.states.SamplerState
-           The sampler state to apply the move to. This is modified.
-
-        See Also
-        --------
-        openmmtools.utils.Timer
-
-        """
-        move_name = self.__class__.__name__
-
-        # Check if we have to use the global cache.
-        if self.context_cache is None:
-            context_cache = cache.global_context_cache
-        else:
-            context_cache = self.context_cache
-
-        # Create integrator.
-        integrator = self._get_integrator(thermodynamic_state)
-
-        # Create context.
-        context, integrator = context_cache.get_context(thermodynamic_state, integrator)
-        thermodynamic_state.apply_to_context(context)
-
-        # If we reassign velocities, we can ignore the ones in sampler_state.
-        sampler_state.apply_to_context(context, ignore_velocities=self.reassign_velocities)
-        if self.reassign_velocities:
-            context.setVelocitiesToTemperature(thermodynamic_state.temperature)
-
-        # Subclasses may implement _before_integration().
-        self._before_integration(context, thermodynamic_state)
-
-        try:
-            nextReport = [None]*len(self.reporters)
-            endStep = self.currentStep + self.n_steps
-            while self.currentStep < endStep:
-                nextSteps = endStep-self.currentStep
-                anyReport = False
-                for i, reporter in enumerate(self.reporters):
-                    nextReport[i] = reporter.describeNextReport(self)
-                    if nextReport[i][0] > 0 and nextReport[i][0] <= nextSteps:
-                        nextSteps = nextReport[i][0]
-                        anyReport = True
-
-                stepsToGo = nextSteps
-                while stepsToGo > 10:
-                    integrator.step(10)
-                    stepsToGo -= 10
-                integrator.step(stepsToGo)
-                self.currentStep += stepsToGo
-
-                if anyReport:
-                    reports = []
-                    context_state = context.getState(
-                         getPositions=True,
-                         getVelocities=True,
-                         getEnergy=True,
-                         enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-                    context_state.currentStep = self.currentStep
-                    context_state.system = thermodynamic_state.get_system()
-
-                    for reporter, report in zip(self.reporters, nextReport):
-                        reports.append((reporter,report))
-                    for reporter, next in reports:
-                         reporter.report(context_state, integrator)
-
-        except Exception:
-            #print(e)
-            traceback.print_exc()
-            # Catches particle positions becoming nan during integration.
-        else:
-            # We get also velocities here even if we don't need them because we
-            # will recycle this State to update the sampler state object. This
-            # way we won't need a second call to Context.getState().
-            context_state = context.getState(
-                getPositions=True,
-                getVelocities=True,
-                getEnergy=True,
-                enforcePeriodicBox=thermodynamic_state.is_periodic)
-
-        # Subclasses can read here info from the context to update internal statistics.
-        self._after_integration(context, thermodynamic_state)
-
-        # Updated sampler state.
-        # This is an optimization around the fact that Collective Variables are not a part of the State,
-        # but are a part of the Context. We do this call twice to minimize duplicating information fetched from
-        # the State.
-        # Update everything but the collective variables from the State object
-        sampler_state.update_from_context(context_state,
-                                          ignore_collective_variables=True)
-        # Update only the collective variables from the Context
-        sampler_state.update_from_context(context,
-                                          ignore_positions=True,
-                                          ignore_velocities=True,
-                                          ignore_collective_variables=False)
-
-
-
-class RandomLigandRotationMove(NCMCMove):
-    def _before_integration(self, context, thermodynamic_state):
-        super(RandomLigandRotationMove, self)._before_integration(context, thermodynamic_state)
-        masses, totalmass = utils.getMasses(self.atom_subset, thermodynamic_state.topology)
-        self.masses = masses
-
-    def _propose_positions(self, positions):
-        """Return new proposed positions.
-
-        These method must be implemented in subclasses.
-
-        Parameters
-        ----------
-        positions : nx3 numpy.ndarray
-            The original positions of the subset of atoms that these move
-            applied to.
-
-        Returns
-        -------
-        proposed_positions : nx3 numpy.ndarray
-            The new proposed positions.
-
-        """
-
-        #print('Proposing positions...')
-        # Calculate the center of mass
-
-        center_of_mass = utils.getCenterOfMass(positions, self.masses)
-        reduced_pos = positions - center_of_mass
-        # Define random rotational move on the ligand
-        rand_quat = mdtraj.utils.uniform_quaternion(size=None)
-        rand_rotation_matrix = mdtraj.utils.rotation_matrix_from_quaternion(rand_quat)
-        #multiply lig coordinates by rot matrix and add back COM translation from origin
-        proposed_positions = numpy.dot(reduced_pos, rand_rotation_matrix) * positions.unit + center_of_mass
-
-        return proposed_positions
 
 ########################
 ## UNDER DEVELOPMENT ###
 ########################
-
 
 # class SideChainMove(Move):
 #     """**NOTE:** Usage of this class requires a valid OpenEye license.
@@ -1305,7 +892,6 @@ class RandomLigandRotationMove(NCMCMove):
 #         old_coord = self._undoBasis(basis_set, center)
 #         adjusted_center = old_coord + particle1
 #         return adjusted_center
-
 
 # class CombinationMove(Move):
 #     """**WARNING:** This class has not been completely tested. Use at your own risk.
