+++ blues/tests/test_ethylene.py
@@ -1,25 +1,19 @@
-import pytest
-import parmed
-import fnmatch
 import logging
-import os
-from blues import utils
-from blues.simulation import SystemFactory, BLUESSampler
-from blues.integrators import AlchemicalExternalLangevinIntegrator
-from blues.moves import RandomLigandRotationMove, ReportLangevinDynamicsMove
-from blues.reporters import (BLUESStateDataStorage, NetCDF4Storage, ReporterConfig, init_logger)
-from blues.settings import Settings
-from simtk import openmm, unit
-from simtk.openmm import app
-from openmmtools.states import SamplerState, ThermodynamicState, CompoundThermodynamicState
-from openmmtools import cache
-import numpy as np
-import mdtraj as md
 from collections import Counter
 
+import mdtraj as md
+import numpy as np
+import parmed
+from openmmtools import cache
+from openmmtools.states import SamplerState, ThermodynamicState
+from simtk import openmm, unit
+
+from blues import utils
+from blues.reporters import BLUESStateDataStorage, NetCDF4Storage
+from blues.ncmc import RandomLigandRotationMove, ReportLangevinDynamicsMove, BLUESSampler
+
 logger = logging.getLogger(__name__)
-#logger = init_logger(logger, level=logging.INFO, stream=True)
-seed = np.random.randint(low=1, high=5000)
+
 
 def runEthyleneTest(dir, N):
     filename = dir.join('ethylene-test_%s' % N)
@@ -33,39 +27,39 @@ def runEthyleneTest(dir, N):
     nIter = 100
     reportInterval = 5
     alchemical_atoms = [2, 3, 4, 5, 6, 7]
-    alchemical_functions = {
-        'lambda_sterics': 'min(1, (1/0.3)*abs(lambda-0.5))',
-        'lambda_electrostatics': 'step(0.2-lambda) - 1/0.2*lambda*step(0.2-lambda) + 1/0.2*(lambda-0.8)*step(lambda-0.8)'
-    }
     context_cache = cache.ContextCache()
 
     # Load a Parmed Structure for the Topology and create our openmm.Simulation
     structure_pdb = utils.get_data_filename('blues', 'tests/data/ethylene_structure.pdb')
     structure = parmed.load_file(structure_pdb)
 
-    nc_reporter = NetCDF4Storage(filename+'_MD.nc', reportInterval)
-    state_reporter = BLUESStateDataStorage(logger, reportInterval,
+    nc_reporter = NetCDF4Storage(filename + '_MD.nc', reportInterval)
+    state_reporter = BLUESStateDataStorage(logger,
+                                           reportInterval,
                                            title='md',
                                            step=True,
                                            speed=True,
-                                           totalSteps=int(n_steps*nIter))
-    nc_reporter1 = NetCDF4Storage(filename+'_NCMC.nc', reportInterval)
-    state_reporter1 = BLUESStateDataStorage(logger, reportInterval,
-                                           title='ncmc',
-                                           step=True,
-                                           speed=True,
-                                           totalSteps=int(n_steps*nIter))
+                                           totalSteps=int(n_steps * nIter))
+    nc_reporter1 = NetCDF4Storage(filename + '_NCMC.nc', reportInterval)
+    state_reporter1 = BLUESStateDataStorage(logger,
+                                            reportInterval,
+                                            title='ncmc',
+                                            step=True,
+                                            speed=True,
+                                            totalSteps=int(n_steps * nIter))
 
     # Iniitialize our Move set
     rot_move = RandomLigandRotationMove(timestep,
-                                         n_steps,
-                                         atom_subset=alchemical_atoms,
-                                         context_cache=context_cache,
-                                       reporters=[nc_reporter1, state_reporter1])
-    langevin_move = ReportLangevinDynamicsMove(timestep, collision_rate, n_steps,
-                                         reassign_velocities=True,
-                                         context_cache=context_cache,
-                                         reporters=[nc_reporter, state_reporter])
+                                        n_steps,
+                                        atom_subset=alchemical_atoms,
+                                        context_cache=context_cache,
+                                        reporters=[nc_reporter1, state_reporter1])
+    langevin_move = ReportLangevinDynamicsMove(timestep,
+                                               collision_rate,
+                                               n_steps,
+                                               reassign_velocities=True,
+                                               context_cache=context_cache,
+                                               reporters=[nc_reporter, state_reporter])
 
     # Load our OpenMM System and create Integrator
     system_xml = utils.get_data_filename('blues', 'tests/data/ethylene_system.xml')
@@ -76,13 +70,13 @@ def runEthyleneTest(dir, N):
     thermodynamic_state = ThermodynamicState(system=system, temperature=temperature)
     sampler_state = SamplerState(positions=structure.positions.in_units_of(unit.nanometers))
 
-    sampler = BLUESSampler(alchemical_atoms,
-                      thermodynamic_state,
-                      sampler_state,
-                      ncmc_move=rot_move,
-                      dynamics_move=langevin_move,
-                      platform=None,
-                      topology=structure.topology)
+    sampler = BLUESSampler(atom_subset=alchemical_atoms,
+                           thermodynamic_state=thermodynamic_state,
+                           sampler_state=sampler_state,
+                           ncmc_move=rot_move,
+                           dynamics_move=langevin_move,
+                           platform=None,
+                           topology=structure.topology)
     sampler.run(nIter)
 
     return filename
@@ -100,7 +94,6 @@ def getPopulations(traj):
 
 
 def graphConvergence(dist, n_points=10):
-    bins = len(dist) / n_points
     bin_count = []
     bin_points = []
     for N in range(1, len(dist) + 1, n_points):
@@ -123,7 +116,7 @@ def graphConvergence(dist, n_points=10):
 
 def test_runEthyleneRepeats(tmpdir):
     dir = tmpdir.mkdir("tmp")
-    outfnames = [runEthyleneTest(dir,N=i) for i in range(5)]
+    outfnames = [runEthyleneTest(dir, N=i) for i in range(5)]
 
     structure_pdb = utils.get_data_filename('blues', 'tests/data/ethylene_structure.pdb')
     trajs = [md.load('%s_MD.nc' % traj, top=structure_pdb) for traj in outfnames]
@@ -142,4 +135,4 @@ def test_runEthyleneRepeats(tmpdir):
     avg_err = np.mean(errs, axis=0)
     print(avg_freq, avg_err, np.absolute(avg_freq - populations))
     check = np.allclose(avg_freq, populations, atol=avg_err)
-    assert check == True
+    assert check is True
